requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
Django==3.1.4
djangorestframework==3.12.2
httpie==2.3.0
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pygments==2.7.3
pylint==2.6.0
PySocks==1.7.1
pytz==2020.4
requests==2.25.0
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.2
wrapt==1.12.1
